Added support for levels 1-5
Changed the limited uses damage value to 0, and moved the others up (makes more sense to me). Also, I fixed a bunch of indentation issues (not to offend you, but I use Eclipse, and not seeing something formatted aggravates my OCD :P). Because I did not write SC, I'm not sure if I missed something in the methods 'addInformation(...)' and 'onItemUse(...)'.