Fix some (Britain-)english names of reinforced blocks (#113)
* Fix some names

Well, as you may or may not know, I am a perfectionist regarding the names of reinforced blocks being equal to the names of their vanilla variants. This pull request started with the Ingot blocks or the former Blocks of Ingot, and then I've found some other minor, minor issues. If there are any problems, please use comments, reviews etc. Thanks!

* Reverting work

Reverted all the GB-english fixes (it seems that there is only 1 fix that hasn't been reverted, lol)

* Update en_gb.json

with the Britain-english texts of my PR.

* Replace tabs at begin of columns with spaces

...so all the lines start with the same position. Sorry, I didn't know that there is a difference between tab and 8 spacebars, because in the editor it looked all the same.

* Remove empty lines