[ENH] Some fixes, some enhancements and two new blocks
This is a collection of enhancements that started ~2 weeks ago. There are many things inside this now, and I don't know if you would appreciate all of them, so please let me know if I have to undo some things. For the exact addings, changes and fixes, look at the changelog please (I hope there are not too many errors in the messages I've written in there).